refactor : atoms refactor
