+++ src/components/atoms/textContent/index.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import './textContent.scss';
-// const ScrollComponent = ({ children, className, scrollActiveLocation } : ScrollProps) =>
+
 interface textProps {
   title: string;
   detail: string;
 }
+
 const TextContent = ({ title, detail }: textProps) => {
   return (
     <div className="text">
